Minor change to this page to redirect the user to the task list
Following yesterday's design review session of the "Create an advert" employer journey, it was pointed out that when selecting to view draft applications, clicking "Edit and submit" for any of the applications would take you to the "as is" page instead of the task list.

It might need some refinement.